Completed task 9. Needs Reviewing

+++ model.js
@@ -65,7 +65,13 @@ exports.updateArticle = (id, votes) => {
 }
 
 exports.removeComment = (id) => {
-   
-    return db.query(`DELETE FROM comments WHERE comment_id = $1 `, [id])
+  return db.query('SELECT * FROM comments WHERE comment_id = $1', [id]).then(({rows})=>{
+    if(rows.length === 0){
+        return Promise.reject({status: 404, msg: "Request Not Found"})
+    } else {
+        return db.query(`DELETE FROM comments WHERE comment_id = $1 `, [id])
+
+    }
+  })
 }
